Add statistics engine parameter
Hive beyond 3.1 (including one coming with CDP7) adds engine field to
statistics entries stored in metastore
(https://issues.apache.org/jira/browse/HIVE-22046). To be able to
communicate with both Hive 3.x and CDP7 we introduce this field to
hive_metastore.thrift as optional. Field will be sent or not by the
calling code depending on which Hive version are we talking to.